Ability To Choose Lesson (only first lesson "body parts" works due to hard-coded curriculum). Code Complete for when data is updated
