+++ IceCreamBuilderMessagesExtension/MessagesViewController.swift
@@ -8,7 +8,10 @@ The root view controller shown by the Messages app.
 import UIKit
 import Messages
 
-class MessagesViewController: MSMessagesAppViewController {
+class MessagesViewController: MSMessagesAppViewController, IceCreamsViewControllerDelegate {
+    func iceCreamsViewControllerDidSelectAdd(_ controller: IceCreamsViewController) {
+    }
+    
 
     // MARK: Properties
     
@@ -55,7 +58,8 @@ class MessagesViewController: MSMessagesAppViewController {
             if iceCream.isComplete {
                 controller = instantiateCompletedIceCreamController(with: iceCream)
             } else {
-                controller = instantiateBuildIceCreamController(with: iceCream)
+                controller = instantiateCompletedIceCreamController(with: iceCream)
+//                controller = instantiateBuildIceCreamController(with: iceCream)
             }
         }
 
@@ -84,17 +88,17 @@ class MessagesViewController: MSMessagesAppViewController {
         return controller
     }
     
-    private func instantiateBuildIceCreamController(with iceCream: IceCream) -> UIViewController {
-        guard let controller = storyboard?.instantiateViewController(withIdentifier: BuildIceCreamViewController.storyboardIdentifier)
-            as? BuildIceCreamViewController
-            else { fatalError("Unable to instantiate a BuildIceCreamViewController from the storyboard") }
-        
-        controller.iceCream = iceCream
-        controller.delegate = self
-        
-        return controller
-    }
-    
+//    private func instantiateBuildIceCreamController(with iceCream: IceCream) -> UIViewController {
+//        guard let controller = storyboard?.instantiateViewController(withIdentifier: BuildIceCreamViewController.storyboardIdentifier)
+//            as? BuildIceCreamViewController
+//            else { fatalError("Unable to instantiate a BuildIceCreamViewController from the storyboard") }
+//        
+//        controller.iceCream = iceCream
+////        controller.delegate = self
+//        
+//        return controller
+//    }
+//    
     private func instantiateCompletedIceCreamController(with iceCream: IceCream) -> UIViewController {
         // Instantiate a `BuildIceCreamViewController`.
         guard let controller = storyboard?.instantiateViewController(withIdentifier: CompletedIceCreamViewController.storyboardIdentifier)
@@ -135,54 +139,54 @@ class MessagesViewController: MSMessagesAppViewController {
 
 /// Extends `MessagesViewController` to conform to the `IceCreamsViewControllerDelegate` protocol.
 
-extension MessagesViewController: IceCreamsViewControllerDelegate {
-
-    func iceCreamsViewControllerDidSelectAdd(_ controller: IceCreamsViewController) {
-        requestPresentationStyle(.expanded)
-    }
-}
-
-/// Extends `MessagesViewController` to conform to the `BuildIceCreamViewControllerDelegate` protocol.
-
-extension MessagesViewController: BuildIceCreamViewControllerDelegate {
-
-    /// - Tag: InsertMessageInConversation
-    func buildIceCreamViewController(_ controller: BuildIceCreamViewController, didSelect iceCreamPart: IceCreamPart) {
-        guard let conversation = activeConversation else { fatalError("Expected a conversation") }
-        guard var iceCream = controller.iceCream else { fatalError("Expected the controller to be displaying an ice cream") }
-
-        // Update the ice cream with the selected body part and determine a caption and description of the change.
-        var messageCaption: String
-        if let base = iceCreamPart as? Base {
-            iceCream.base = base
-            messageCaption = NSLocalizedString("Let's build an ice cream", comment: "")
-        } else if let scoops = iceCreamPart as? Scoops {
-            iceCream.scoops = scoops
-            messageCaption = NSLocalizedString("I added some scoops", comment: "")
-        } else if let topping = iceCreamPart as? Topping {
-            iceCream.topping = topping
-            messageCaption = NSLocalizedString("Our finished ice cream", comment: "")
-        } else {
-            fatalError("Unexpected type of ice cream part selected.")
-        }
-
-        // Create a new message with the same session as any currently selected message.
-        let message = composeMessage(with: iceCream, caption: messageCaption, session: conversation.selectedMessage?.session)
-
-        // Add the message to the conversation.
-        conversation.insert(message) { error in
-            if let error = error {
-                print(error)
-            }
-        }
-
-        // If the ice cream is complete, save it in the history.
-        if iceCream.isComplete {
-            var history = IceCreamHistory.load()
-            history.append(iceCream)
-            history.save()
-        }
-        
-        dismiss()
-    }
-}
+//extension MessagesViewController: IceCreamsViewControllerDelegate {
+//
+//    func iceCreamsViewControllerDidSelectAdd(_ controller: IceCreamsViewController) {
+//        requestPresentationStyle(.expanded)
+//    }
+//}
+//
+///// Extends `MessagesViewController` to conform to the `BuildIceCreamViewControllerDelegate` protocol.
+//
+//extension MessagesViewController: BuildIceCreamViewControllerDelegate {
+//
+//    /// - Tag: InsertMessageInConversation
+//    func buildIceCreamViewController(_ controller: BuildIceCreamViewController, didSelect iceCreamPart: IceCreamPart) {
+//        guard let conversation = activeConversation else { fatalError("Expected a conversation") }
+//        guard var iceCream = controller.iceCream else { fatalError("Expected the controller to be displaying an ice cream") }
+//
+//        // Update the ice cream with the selected body part and determine a caption and description of the change.
+//        var messageCaption: String
+//        if let base = iceCreamPart as? Base {
+//            iceCream.base = base
+//            messageCaption = NSLocalizedString("Let's build an ice cream", comment: "")
+//        } else if let scoops = iceCreamPart as? Scoops {
+//            iceCream.scoops = scoops
+//            messageCaption = NSLocalizedString("I added some scoops", comment: "")
+//        } else if let topping = iceCreamPart as? Topping {
+//            iceCream.topping = topping
+//            messageCaption = NSLocalizedString("Our finished ice cream", comment: "")
+//        } else {
+//            fatalError("Unexpected type of ice cream part selected.")
+//        }
+//
+//        // Create a new message with the same session as any currently selected message.
+//        let message = composeMessage(with: iceCream, caption: messageCaption, session: conversation.selectedMessage?.session)
+//
+//        // Add the message to the conversation.
+//        conversation.insert(message) { error in
+//            if let error = error {
+//                print(error)
+//            }
+//        }
+//
+//        // If the ice cream is complete, save it in the history.
+//        if iceCream.isComplete {
+//            var history = IceCreamHistory.load()
+//            history.append(iceCream)
+//            history.save()
+//        }
+//        
+//        dismiss()
+//    }
+//}
